requirements.txt
# ------------------------
# Core Framework
# ------------------------
Django>=5.0,<6.0
python-dotenv>=1.0.1

# ------------------------
# Database (use PostgreSQL in production)
# ------------------------
psycopg2-binary>=2.9.9

# ------------------------
# PDF Parsing
# ------------------------
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
tldextract>=5.1.1

# ------------------------
# Email Sending (Gmail API OAuth)
# ------------------------
google-api-python-client>=2.118.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# ------------------------
# Background Tasks
# ------------------------
celery>=5.3.6
redis>=5.0.1

# ------------------------
# File Handling
# ------------------------
django-cleanup>=9.0.0
Pillow>=10.2.0

# ------------------------
# Utilities
# ------------------------
django-widget-tweaks>=1.5.0
requests>=2.31.0